moved task and tasks logic into a store
